Submit Changes
Form labels now change color depending on which action is being
performed. Red indicates the field must be filled and black indicates
optional

Changed form submission requirements. Adding an organization requires
only a shortname and a name, editing  or deleting an organization
requires the dropdown short name or name dropdown be selected. Adding a
platform requires the name, description, lat, and lon be filled, and
editing or deleting a platform requires that a organization and
platform have been selected. Adding a sensor requires the type and
vertical position be filled, and editing and deleting a sensor require
the sensor type be selected.

Currently, editing and deleting a sensor do not work properly.